Preliminary #wsyn operator changes.

+++ indri/branches/strohman/include/indri/DirichletTermScoreFunction.hpp
@@ -34,12 +34,12 @@ public:
     _muTimesCollectionFrequency = _mu * _collectionFrequency;
   }
 
-  double scoreOccurrence( int occurrences, int contextSize ) {
+  double scoreOccurrence( double occurrences, int contextSize ) {
     double seen = ( double(occurrences) + _muTimesCollectionFrequency ) / ( double(contextSize) + _mu );
     return log( seen );
   }
 
-  double scoreOccurrence( int occurrences, int contextSize, int documentOccurrences, int documentLength ) {
+  double scoreOccurrence( double occurrences, int contextSize, double documentOccurrences, int documentLength ) {
     // can't two-level smooth with dirichlet
     return scoreOccurrence( occurrences, contextSize );
   }
