+++ indri/branches/strohman/include/indri/ListBeliefNode.hpp
@@ -41,8 +41,8 @@ private:
 private:
   // computes the length of the scored context
   inline int _contextLength( int begin, int end );
-  inline int _contextOccurrences( int begin, int end );
-  inline int _documentOccurrences();
+  inline double _contextOccurrences( int begin, int end );
+  inline double _documentOccurrences();
 
 public:
   ListBeliefNode( const std::string& name,
